# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: EthixAI API
  description: |
    Ethical AI Governance Platform API - Real-time bias detection, SHAP explanations,
    and comprehensive monitoring for financial decision-making systems.
  version: '1.0.0'
  contact:
    name: EthixAI Support
    email: support@ethixai.com
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://api.ethixai.com
    description: Production
...
paths:
...
  /api/v1/models/{id}/trigger-retrain:
    post:
      tags: [Models]
      summary: Trigger model retraining
      operationId: triggerRetrain
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Model ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  minLength: 3
                baseline_snapshot_id:
                  type: string
                notes:
                  type: string
      responses:
        '202':
          description: Retraining queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: queued
                  requestId:
                    type: string
        '403':
          description: Admin role required
  /api/v1/retrain/{requestId}/complete:
    post:
      tags: [Models]
      summary: Report retraining completion (CI callback)
      operationId: completeRetrain
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                performance_metrics:
                  type: object
      responses:
        '200':
          description: Retraining marked complete
        '401':
          description: Invalid callback secret
        '404':
          description: Request not found
  /api/v1/models/{id}/versions:
    get:
      tags: [Models]
      summary: List model versions
      operationId: listModelVersions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of model versions
          content:
            application/json:
              schema:
                type: object
                properties:
                  versions:
                    type: array
                    items:
                      $ref: '#/components/schemas/ModelVersion'
  /api/v1/models/{id}/versions/{version}/promote:
    post:
      tags: [Models]
      summary: Promote model version to production
      operationId: promoteModel
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: version
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Model promoted
        '403':
          description: Admin role required

  # ==================== EVALUATION ====================
...
  /api/policies:
    get:
      tags: [Policies]
      summary: List policies
      operationId: listPolicies
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of policies
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Policy'

    post:
      tags: [Policies]
      summary: Create policy
      operationId: createPolicy
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyInput'
      responses:
        '201':
          description: Policy created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
  /api/policies/{policyId}:
    get:
      tags: [Policies]
      summary: Get policy details
      operationId: getPolicy
      security:
        - bearerAuth: []
      parameters:
        - name: policyId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Policy details
        '404':
          description: Policy not found

    put:
      tags: [Policies]
      summary: Update policy
      operationId: updatePolicy
      security:
        - bearerAuth: []
      parameters:
        - name: policyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyInput'
      responses:
        '200':
          description: Policy updated
        '404':
          description: Policy not found

    delete:
      tags: [Policies]
      summary: Delete policy
      operationId: deletePolicy
      security:
        - bearerAuth: []
      parameters:
        - name: policyId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Policy deleted
        '404':
          description: Policy not found

  # ==================== REPORTS ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ModelVersion:
      type: object
      properties:
        version:
          type: string
        status:
          type: string
          enum: [draft, staging, production, archived]
        created_at:
          type: string
          format: date-time
        performance_metrics:
          type: object

    # ==================== GOVERNANCE SCHEMAS ====================
...
    PolicyInput:
      type: object
      required: [name, rules]
      properties:
        name:
          type: string
        description:
          type: string
        rules:
          type: array
          items:
            type: object
            properties:
              metric:
                type: string
              threshold:
                type: number
              operator:
                type: string
                enum: [gt, gte, lt, lte, eq, neq]
        enforcement:
          type: string
          enum: [block, warn, log]
        enabled:
          type: boolean
          default: true
    Policy:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        description:
          type: string
        rules:
          type: array
        enforcement:
          type: string
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # ==================== REPORT SCHEMAS ====================
...